+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Change Log
 
+## [0.2.51] - 2018-05-01
+### Changed
+- The `advocate` privilege grants access to see user information using
+  the API.
+- The `interview_list()` function and `/api/interviews` API now act
+  upon interviews of anonymous users.
+### Fixed
+- Questions not centered on screen when navigation bar in use but empty.
+- Google Places API populating `unit` with `subpremise` and leaving
+  `city` blank.
+
 ## [0.2.50] - 2018-04-28
 ### Changed
 - Upgraded user interface from Bootstrap 3.3 to Bootstrap 4.0.0.